pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "borehole-to-surfer"
version = "0.1.0"
description = "Windows desktop tool for generating Surfer-ready borehole section outputs."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "numpy>=1.26.0",
    "pandas>=2.2.0",
    "pyshp>=2.3.1",
    "Pillow>=10.4.0",
    "pyproj>=3.7.0",
]

[tool.setuptools.packages.find]
where = ["src"]
